Remove class `active` from both pointers.
When moving the lower pointer to the right side of the upper pointer or the
other way round (moving the upper pointer to the left side of the lower
pointer), the class active is not remove from the correct pointer.

Removing the class `active` form both pointers fixes the problem.

Fixes #45.